[Refactor] move `INotificationService` labels into a new file to avoid env issues
