+++ src/workbench/services/theme/themeService.ts
@@ -18,9 +18,10 @@ import { WorkbenchConfiguration } from "src/workbench/services/workbench/configu
 import { mixin } from "src/base/common/utilities/object";
 import { Color, ColorMap } from "src/base/common/color";
 import { noop } from "src/base/common/performance";
-import { INotificationService, NotificationTypes } from "src/workbench/services/notification/notificationService";
+import { NotificationTypes } from "src/workbench/services/notification/notificationService";
 import { ColorRegistrant } from "src/workbench/services/theme/colorRegistrant";
 import { Strings } from "src/base/common/utilities/string";
+import { INotificationService } from "src/workbench/services/notification/notification";
 
 export const IThemeService = createService<IThemeService>('theme-service');
