+++ src/workbench/services/workbench/command.register.ts
@@ -5,7 +5,6 @@ import { RegistrantType, createRegister } from "src/platform/registrant/common/r
 import { FileCommands } from "src/workbench/services/fileTree/fileCommands";
 import { Command } from "src/platform/command/common/command";
 import { IServiceProvider } from "src/platform/instantiation/common/instantiation";
-import { INotificationService } from "src/workbench/services/notification/notificationService";
 import { errorToMessage } from "src/base/common/utilities/panic";
 import { ILogService } from "src/base/common/logger";
 import { IBrowserZoomService } from "src/workbench/services/zoom/zoomService";
@@ -18,6 +17,7 @@ import { IBrowserInspectorService } from "src/platform/inspector/common/inspecto
 import { INavigationViewService } from "src/workbench/parts/navigationPanel/navigationView/navigationView";
 import { ExplorerView } from "src/workbench/contrib/explorer/explorer";
 import { IRecentOpenService } from "src/platform/app/browser/recentOpenService";
+import { INotificationService } from "src/workbench/services/notification/notification";
 
 export const rendererWorkbenchCommandRegister = createRegister(
     RegistrantType.Command,
